Fix very cool bug
This was the reason hypixel & co didn't work! Because the IPv4 of the server was put in the Address field of the Handshake packet, and not the real domain name, the address. Hence, the server couldn't properly do its logic given the address! But it's fixed!!! That was such a quick and simple fix, but with such consequences to it